[fuzzing] jellyfish fuzzers
Closes: #5936
