+++ storage/jellyfish-merkle/Cargo.toml
@@ -17,6 +17,7 @@ num-derive = "0.3.2"
 num-traits = "0.2.12"
 proptest = { version = "0.10.1", optional = true }
 proptest-derive = { version = "0.2.0", optional = true }
+rand = {version = "0.7.3", optional = true }
 serde = { version = "1.0.115", features = ["derive"] }
 thiserror = "1.0.20"
 
@@ -38,4 +39,4 @@ libra-types = { path = "../../types", version = "0.1.0", features = ["fuzzing"] 
 
 [features]
 default = []
-fuzzing = ["proptest", "proptest-derive", "libra-crypto/fuzzing", "libra-types/fuzzing", "libra-nibble/fuzzing"]
+fuzzing = ["proptest", "rand", "proptest-derive", "libra-crypto/fuzzing", "libra-types/fuzzing", "libra-nibble/fuzzing"]
