+++ testsuite/libra-fuzzer/Cargo.toml
@@ -32,6 +32,7 @@ executor = { path = "../../execution/executor", version = "0.1.0", features = ["
 executor-types = { path = "../../execution/executor-types", version = "0.1.0", features = ["fuzzing"] }
 language-e2e-tests = { path = "../../language/testing-infra/e2e-tests", version = "0.1.0" }
 libra-crypto = { path = "../../crypto/crypto", version = "0.1.0", features = ["fuzzing"]}
+libra-jellyfish-merkle = { path = "../../storage/jellyfish-merkle", version = "0.1.0", features = ["fuzzing"] }
 libra-json-rpc = { path = "../../json-rpc", version = "0.1.0", features = ["fuzzing"] }
 libra-mempool = { path = "../../mempool", version = "0.1.0" }
 libra-secure-json-rpc = { path = "../../secure/json-rpc", version = "0.1.0", features = ["fuzzing"]}
